[Added] - moved files into folders and seperated functions as per file basis
