+++ src/constants.js
@@ -1,0 +1,9 @@
+/**
+ * Constants:
+ * Mean Radius of Earth.
+ * PI/360.
+ */
+const R = 6378.137;
+const PI_360 = Math.PI / 360;
+
+module.exports = { R, PI_360 };
